+++ src/components/Profile/Profile.js
@@ -1,0 +1,65 @@
+import React, { useState } from 'react'
+import DisplayCard from '../DisplayCard'
+import styles from './Profile.module.css'
+import CSSModuleMapper from '../../utils/cssModuleMapper'
+
+const css = CSSModuleMapper(styles)
+
+const Profile = () => {
+  const [data, setData] = useState({})
+  const [username, setUsername] = useState('')
+  const [repositories, setRepositories] = useState([])
+
+  const onChangeHandler = (e) => {
+    setUsername(e.target.value)
+  }
+
+  const submitHandler = async (e) => {
+    e.preventDefault()
+
+    const profile = await fetch(`https://api.github.com/users/${username}`)
+    const profileJson = await profile.json()
+    console.log(profileJson)
+
+    const repositories = await fetch(profileJson.repos_url)
+    const repoJson = await repositories.json()
+    console.log(repoJson)
+
+    if (profileJson) {
+      setData(profileJson)
+      setRepositories(repoJson)
+    }
+  }
+  return (
+    <>
+      <div className={css('page')}>
+        <div className="ui search">
+          <div className="ui icon input">
+            <i className="search icon"></i>
+            <input
+              className="prompt"
+              placeholder="Whom you wanna stalk?"
+              type="text"
+              value={username}
+              onChange={onChangeHandler}
+            />
+          </div>
+
+          <button
+            className={`ui violet button ${css('go-btn')}`}
+            type="submit"
+            onClick={submitHandler}
+          >
+            <i className="github icon"></i>
+            Go ahead!
+          </button>
+
+          <center>
+            <DisplayCard data={data} repositories={repositories} />
+          </center>
+        </div>
+      </div>
+    </>
+  )
+}
+export default Profile
