+++ src/components/Profile/Profile.module.css
@@ -1,0 +1,7 @@
+.page {
+  padding: 20px;
+}
+
+.go-btn {
+  margin-left: 20px !important;
+}
