+++ src/components/Profile/index.js
@@ -1,0 +1,5 @@
+/**
+ * Re-export to prevent nested path
+ */
+import Profile from './Profile'
+export default Profile
